Modify class ImageState to represent current state of image
Add methods getWidth(), getHeight(), getBuffer that returns
respectively width, height and bytes buffer of image and method
getPixel() which returns pixel with x and y coordinates and
throws an error if x or y have wrong values.

Add method setBuffer() which overwrites the bytes buffer with
other given buffer and method setPixel() which sets new value to pixel
with x and y coordinates and throws an error if x or y have wrong values.

To prove that above methods work properly, modify Project::main()
to print results of using them.