Redesign Transform API to suit Areas and Threads
Now classes implementing Transform don't override apply()
which is final and public API, instead they override
methods that define output dimensions and getPixel()
which returns a single pixel with coords x,y based on input.

Define class Effect which is a special case of Transform,
that does not change dimensions of ImageState. Effect makes
getOutputHeight and getOutputWidth final.

Thanks to per-pixel implementation, it will be easier to
add Multithreading and Area support to Transform.