Allow for zooming in on JRenderer
Implement a zoom feature on JRenderer using keyPress
and keyRelease event listeners on parent JPanel and
a mouseScrollWheel listener. Allow for 4 levels of
linear zoom (each pixel takes 1, 2, 3 or 4 pixels
on the screen).

KeyPress and keyRelease events are needed to only
allow zooming when a certain key is pressed, currently
set to shift. The key state is tracked by a zoomKeyPressed
private boolean variable.